Moves {Ecdsa,Rsa}PublicKey classes into their own modules
"Tidier, each file [is] shorter and with narrower scope." - eranm
