+++ python/ct/crypto/verify.py
@@ -1,23 +1,18 @@
 """Verify CT log statements."""
 
-import hashlib
 import io
 import struct
 
+from ct.crypto import ecdsa_keys
 from ct.crypto import error
 from ct.crypto import merkle
-from ct.crypto import pem
+from ct.crypto import rsa_keys
 from ct.crypto.asn1 import oid
 from ct.crypto.asn1 import x509_extension as x509_ext
 from ct.crypto.asn1 import x509_name
 from ct.proto import client_pb2
 from ct.proto import ct_pb2
 from ct.serialization import tls_message
-
-import Crypto.Hash.SHA256
-import Crypto.PublicKey.RSA
-import Crypto.Signature.PKCS1_v1_5
-import ecdsa
 
 SUPPORTED_HASH_ALGORITHMS = (
     ct_pb2.DigitallySigned.SHA256,
@@ -191,142 +186,6 @@ def _create_dst_entry(sct, chain):
 
     return entry
 
-class RsaPublicKey(object):
-    """Wraps an RSA public key."""
-
-    """The signature algorithm used for this public key."""
-    SIGNATURE_ALGORITHM = client_pb2.DigitallySigned.RSA
-    """The hash algorithm used for this public key."""
-    HASH_ALGORITHM = client_pb2.DigitallySigned.SHA256
-
-    """Markers to look for when reading a PEM-encoded RSA public key."""
-    __READ_MARKERS = ("PUBLIC KEY", "RSA PUBLIC KEY")
-    """A marker to write when writing a PEM-encoded RSA public key."""
-    __WRITE_MARKER = "RSA PUBLIC KEY"
-
-    def __init__(self, key_info):
-        """Loads a PEM-encoded RSA public key.
-
-        Args:
-        - key_info: KeyInfo protobuf message
-
-        Raises:
-        - PemError: If the key has an invalid encoding
-        """
-        if (key_info.type != client_pb2.KeyInfo.RSA):
-            raise error.UnsupportedAlgorithmError(
-                "Expected RSA key, but got key type %d" % key_info.type)
-
-        # Will raise a PemError on invalid encoding
-        self.__der, _ = pem.from_pem(key_info.pem_key, self.__READ_MARKERS)
-        try:
-            self.__key = Crypto.PublicKey.RSA.importKey(self.__der)
-        except (ValueError, IndexError, TypeError) as e:
-            raise error.EncodingError(e)
-
-    def __repr__(self):
-        return "%r(public key: %r)" % (self.__class__.__name__,
-                                       pem.to_pem(self.__der,
-                                                  self.__WRITE_MARKER))
-
-    def __str__(self):
-        return "%s(public key: %s)" % (self.__class__.__name__,
-                                       pem.to_pem(self.__der,
-                                                  self.__WRITE_MARKER))
-
-    @error.returns_true_or_raises
-    def verify(self, signature_input, signature):
-        """Verifies the signature was created by the owner of this key.
-
-        Args:
-        - signature_input: The data that was originally signed.
-        - signature: An RSA SHA256 signature.
-
-        Returns:
-        - True if the signature verifies.
-
-        Raises:
-        - error.SignatureError: If the signature fails verification.
-        """
-        verifier = Crypto.Signature.PKCS1_v1_5.new(self.__key)
-        hash = Crypto.Hash.SHA256.new(signature_input)
-
-        if verifier.verify(hash, signature):
-            return True
-        else:
-            raise error.SignatureError("Signature did not verify: %s",
-                                       signature.encode("hex"))
-
-
-class EcdsaPublicKey(object):
-    """Wraps an ECDSA public key."""
-
-    """The signature algorithm used for this public key."""
-    SIGNATURE_ALGORITHM = client_pb2.DigitallySigned.ECDSA
-    """The hash algorithm used for this public key."""
-    HASH_ALGORITHM = client_pb2.DigitallySigned.SHA256
-
-    """Markers to look for when reading a PEM-encoded ECDSA public key."""
-    __READ_MARKERS = ("PUBLIC KEY", "ECDSA PUBLIC KEY")
-    """A marker to write when writing a PEM-encoded ECDSA public key."""
-    __WRITE_MARKER = "ECDSA PUBLIC KEY"
-
-    def __init__(self, key_info):
-        """Loads a PEM-encoded ECDSA public key.
-
-        Args:
-        - key_info: KeyInfo protobuf message
-
-        Raises:
-        - PemError: If the key has an invalid encoding
-        """
-        if (key_info.type != client_pb2.KeyInfo.ECDSA):
-            raise error.UnsupportedAlgorithmError(
-                "Expected ECDSA key, but got key type %d" % key_info.type)
-
-        # Will raise a PemError on invalid encoding
-        self.__der, _ = pem.from_pem(key_info.pem_key, self.__READ_MARKERS)
-        try:
-            self.__key = ecdsa.VerifyingKey.from_der(self.__der)
-        except ecdsa.der.UnexpectedDER as e:
-            raise error.EncodingError(e)
-
-    def __repr__(self):
-        return "%r(public key: %r)" % (self.__class__.__name__,
-                                       pem.to_pem(self.__der,
-                                                  self.__WRITE_MARKER))
-
-    def __str__(self):
-        return "%s(public key: %s)" % (self.__class__.__name__,
-                                       pem.to_pem(self.__der,
-                                                  self.__WRITE_MARKER))
-
-    @error.returns_true_or_raises
-    def verify(self, signature_input, signature):
-        """Verifies the signature was created by the owner of this key.
-
-        Args:
-        - signature_input: The data that was originally signed.
-        - signature: An ECDSA SHA256 signature.
-
-        Returns:
-        - True if the signature verifies.
-
-        Raises:
-        - error.EncodingError: If the signature encoding is invalid.
-        - error.SignatureError: If the signature fails verification.
-        """
-        try:
-            return self.__key.verify(signature, signature_input,
-                                     hashfunc=hashlib.sha256,
-                                     sigdecode=ecdsa.util.sigdecode_der)
-        except ecdsa.der.UnexpectedDER:
-            raise error.EncodingError("Invalid DER encoding for signature %s",
-                                      signature.encode("hex"))
-        except ecdsa.keys.BadSignatureError:
-            raise error.SignatureError("Signature did not verify: %s",
-                                       signature.encode("hex"))
-
 
 class LogVerifier(object):
     """CT log verifier."""
@@ -335,9 +194,9 @@ class LogVerifier(object):
         """Initialize from KeyInfo protocol buffer and a MerkleVerifier."""
         self.__merkle_verifier = merkle_verifier
         if (key_info.type == client_pb2.KeyInfo.ECDSA):
-            self.__pubkey = EcdsaPublicKey(key_info)
+            self.__pubkey = ecdsa_keys.EcdsaPublicKey(key_info)
         elif (key_info.type == client_pb2.KeyInfo.RSA):
-            self.__pubkey = RsaPublicKey(key_info)
+            self.__pubkey = rsa_keys.RsaPublicKey(key_info)
         else:
             raise error.UnsupportedAlgorithmError("Key type %d not supported" %
                                                   key_info.type)
